requirements.txt
py-cord==2.4.1
python-dotenv==0.21.1
jsonobject==2.1.0
